Separated TextEditorComponent sub-components
Each now have its own file. It's way easier to understand that way,
even considering that the editor still is "too big"
