+++ src/text-editor/component.js
@@ -1,23 +1,32 @@
 /* global ResizeObserver */
 
 const etch = require('etch');
+const $ = etch.dom;
 const { Point, Range } = require('text-buffer');
 const LineTopIndex = require('line-top-index');
-const TextEditor = require('../text-editor');
-const { isPairedCharacter } = require('./utils');
+const { debounce } = require('underscore');
 const electron = require('electron');
 const clipboard = electron.clipboard;
-const $ = etch.dom;
 
-let TextEditorElement;
+const TextEditor = require('../text-editor');
+const TextEditorElement = require('./element');;
+const GutterContainerComponent = require('./gutter-container-component');
+const CursorsAndInputComponent = require('./cursors-and-input-component');
+const LinesTileComponent = require('./lines-tile-component');
+const HighlightsComponent = require('./highlights-component');
+const DummyScrollbarComponent = require('./dummy-scrollbar-component');
+const OverlayComponent = require('./overlay-component');
+const NodePool = require('./node-pool');
+const {
+  isPairedCharacter, roundToPhysicalPixelBoundary, ceilToPhysicalPixelBoundary,
+  constrainRangeToRows
+} = require('./utils');
 
 const DEFAULT_ROWS_PER_TILE = 6;
 const NORMAL_WIDTH_CHARACTER = 'x';
 const DOUBLE_WIDTH_CHARACTER = '我';
 const HALF_WIDTH_CHARACTER = 'ﾊ';
 const KOREAN_CHARACTER = '세';
-const NBSP_CHARACTER = '\u00a0';
-const ZERO_WIDTH_NBSP_CHARACTER = '\ufeff';
 const MOUSE_DRAG_AUTOSCROLL_MARGIN = 40;
 const CURSOR_BLINK_RESUME_DELAY = 300;
 const CURSOR_BLINK_PERIOD = 800;
@@ -65,8 +74,6 @@ module.exports = class TextEditorComponent {
     if (props.element) {
       this.element = props.element;
     } else {
-      if (!TextEditorElement)
-        TextEditorElement = require('./element');
       this.element = TextEditorElement.createTextEditorElement();
     }
     this.element.initialize(this);
@@ -1881,7 +1888,7 @@ module.exports = class TextEditorComponent {
     this.lastKeydown = event;
   }
 
-  didKeypress(event) {
+  didKeypress(_) {
     this.lastKeydownBeforeKeypress = this.lastKeydown;
 
     // This cancels the accented character behavior if we type a key normally
@@ -3396,1649 +3403,7 @@ module.exports = class TextEditorComponent {
   getPlatform() {
     return this.props.platform || process.platform;
   }
-
-  getChromeVersion() {
-    return this.props.chromeVersion || parseInt(process.versions.chrome);
-  }
 };
-
-class DummyScrollbarComponent {
-  constructor(props) {
-    this.props = props;
-    etch.initialize(this);
-  }
-
-  update(newProps) {
-    const oldProps = this.props;
-    this.props = newProps;
-    etch.updateSync(this);
-
-    const shouldFlushScrollPosition =
-      newProps.scrollTop !== oldProps.scrollTop ||
-      newProps.scrollLeft !== oldProps.scrollLeft;
-    if (shouldFlushScrollPosition) this.flushScrollPosition();
-  }
-
-  flushScrollPosition() {
-    if (this.props.orientation === 'horizontal') {
-      this.element.scrollLeft = this.props.scrollLeft;
-    } else {
-      this.element.scrollTop = this.props.scrollTop;
-    }
-  }
-
-  render() {
-    const {
-      orientation,
-      scrollWidth,
-      scrollHeight,
-      verticalScrollbarWidth,
-      horizontalScrollbarHeight,
-      canScroll,
-      forceScrollbarVisible,
-      didScroll
-    } = this.props;
-
-    const outerStyle = {
-      position: 'absolute',
-      contain: 'content',
-      zIndex: 1,
-      willChange: 'transform'
-    };
-    if (!canScroll) outerStyle.visibility = 'hidden';
-
-    const innerStyle = {};
-    if (orientation === 'horizontal') {
-      let right = verticalScrollbarWidth || 0;
-      outerStyle.bottom = 0;
-      outerStyle.left = 0;
-      outerStyle.right = right + 'px';
-      outerStyle.height = '15px';
-      outerStyle.overflowY = 'hidden';
-      outerStyle.overflowX = forceScrollbarVisible ? 'scroll' : 'auto';
-      outerStyle.cursor = 'default';
-      innerStyle.height = '15px';
-      innerStyle.width = (scrollWidth || 0) + 'px';
-    } else {
-      let bottom = horizontalScrollbarHeight || 0;
-      outerStyle.right = 0;
-      outerStyle.top = 0;
-      outerStyle.bottom = bottom + 'px';
-      outerStyle.width = '15px';
-      outerStyle.overflowX = 'hidden';
-      outerStyle.overflowY = forceScrollbarVisible ? 'scroll' : 'auto';
-      outerStyle.cursor = 'default';
-      innerStyle.width = '15px';
-      innerStyle.height = (scrollHeight || 0) + 'px';
-    }
-
-    return $.div(
-      {
-        className: `${orientation}-scrollbar`,
-        style: outerStyle,
-        on: {
-          scroll: didScroll,
-          mousedown: this.didMouseDown
-        }
-      },
-      $.div({ style: innerStyle })
-    );
-  }
-
-  didMouseDown(event) {
-    let { bottom, right } = this.element.getBoundingClientRect();
-    const clickedOnScrollbar =
-      this.props.orientation === 'horizontal'
-        ? event.clientY >= bottom - this.getRealScrollbarHeight()
-        : event.clientX >= right - this.getRealScrollbarWidth();
-    if (!clickedOnScrollbar) this.props.didMouseDown(event);
-  }
-
-  getRealScrollbarWidth() {
-    return this.element.offsetWidth - this.element.clientWidth;
-  }
-
-  getRealScrollbarHeight() {
-    return this.element.offsetHeight - this.element.clientHeight;
-  }
-}
-
-class GutterContainerComponent {
-  constructor(props) {
-    this.props = props;
-    etch.initialize(this);
-  }
-
-  update(props) {
-    if (this.shouldUpdate(props)) {
-      this.props = props;
-      etch.updateSync(this);
-    }
-  }
-
-  shouldUpdate(props) {
-    return (
-      !props.measuredContent ||
-      props.lineNumberGutterWidth !== this.props.lineNumberGutterWidth
-    );
-  }
-
-  render() {
-    const {
-      hasInitialMeasurements,
-      scrollTop,
-      scrollHeight,
-      guttersToRender,
-      decorationsToRender
-    } = this.props;
-
-    const innerStyle = {
-      willChange: 'transform',
-      display: 'flex'
-    };
-
-    if (hasInitialMeasurements) {
-      innerStyle.transform = `translateY(${-roundToPhysicalPixelBoundary(
-        scrollTop
-      )}px)`;
-    }
-
-    return $.div(
-      {
-        ref: 'gutterContainer',
-        key: 'gutterContainer',
-        className: 'gutter-container',
-        style: {
-          position: 'relative',
-          zIndex: 1,
-          backgroundColor: 'inherit'
-        }
-      },
-      $.div(
-        { style: innerStyle },
-        guttersToRender.map(gutter => {
-          if (gutter.type === 'line-number') {
-            return this.renderLineNumberGutter(gutter);
-          } else {
-            return $(CustomGutterComponent, {
-              key: gutter,
-              element: gutter.getElement(),
-              name: gutter.name,
-              visible: gutter.isVisible(),
-              height: scrollHeight,
-              decorations: decorationsToRender.customGutter.get(gutter.name)
-            });
-          }
-        })
-      )
-    );
-  }
-
-  renderLineNumberGutter(gutter) {
-    const {
-      rootComponent,
-      showLineNumbers,
-      hasInitialMeasurements,
-      lineNumbersToRender,
-      renderedStartRow,
-      renderedEndRow,
-      rowsPerTile,
-      decorationsToRender,
-      didMeasureVisibleBlockDecoration,
-      scrollHeight,
-      lineNumberGutterWidth,
-      lineHeight
-    } = this.props;
-
-    if (!gutter.isVisible()) {
-      return null;
-    }
-
-    const oneTrueLineNumberGutter = gutter.name === 'line-number';
-    const ref = oneTrueLineNumberGutter ? 'lineNumberGutter' : undefined;
-    const width = oneTrueLineNumberGutter ? lineNumberGutterWidth : undefined;
-
-    if (hasInitialMeasurements) {
-      const {
-        maxDigits,
-        keys,
-        bufferRows,
-        screenRows,
-        softWrappedFlags,
-        foldableFlags
-      } = lineNumbersToRender;
-      return $(LineNumberGutterComponent, {
-        ref,
-        element: gutter.getElement(),
-        name: gutter.name,
-        className: gutter.className,
-        labelFn: gutter.labelFn,
-        onMouseDown: gutter.onMouseDown,
-        onMouseMove: gutter.onMouseMove,
-        rootComponent: rootComponent,
-        startRow: renderedStartRow,
-        endRow: renderedEndRow,
-        rowsPerTile: rowsPerTile,
-        maxDigits: maxDigits,
-        keys: keys,
-        bufferRows: bufferRows,
-        screenRows: screenRows,
-        softWrappedFlags: softWrappedFlags,
-        foldableFlags: foldableFlags,
-        decorations: decorationsToRender.lineNumbers.get(gutter.name) || [],
-        blockDecorations: decorationsToRender.blocks,
-        didMeasureVisibleBlockDecoration: didMeasureVisibleBlockDecoration,
-        height: scrollHeight,
-        width,
-        lineHeight: lineHeight,
-        showLineNumbers
-      });
-    } else {
-      return $(LineNumberGutterComponent, {
-        ref,
-        element: gutter.getElement(),
-        name: gutter.name,
-        className: gutter.className,
-        onMouseDown: gutter.onMouseDown,
-        onMouseMove: gutter.onMouseMove,
-        maxDigits: lineNumbersToRender.maxDigits,
-        showLineNumbers
-      });
-    }
-  }
-}
-
-class LineNumberGutterComponent {
-  constructor(props) {
-    this.props = props;
-    this.element = this.props.element;
-    this.virtualNode = $.div(null);
-    this.virtualNode.domNode = this.element;
-    this.nodePool = new NodePool();
-    etch.updateSync(this);
-  }
-
-  update(newProps) {
-    if (this.shouldUpdate(newProps)) {
-      this.props = newProps;
-      etch.updateSync(this);
-    }
-  }
-
-  render() {
-    const {
-      rootComponent,
-      showLineNumbers,
-      height,
-      width,
-      startRow,
-      endRow,
-      rowsPerTile,
-      maxDigits,
-      keys,
-      bufferRows,
-      screenRows,
-      softWrappedFlags,
-      foldableFlags,
-      decorations,
-      className
-    } = this.props;
-
-    let children = null;
-
-    if (bufferRows) {
-      children = new Array(rootComponent.renderedTileStartRows.length);
-      for (let i = 0; i < rootComponent.renderedTileStartRows.length; i++) {
-        const tileStartRow = rootComponent.renderedTileStartRows[i];
-        const tileEndRow = Math.min(endRow, tileStartRow + rowsPerTile);
-        const tileChildren = new Array(tileEndRow - tileStartRow);
-        for (let row = tileStartRow; row < tileEndRow; row++) {
-          const indexInTile = row - tileStartRow;
-          const j = row - startRow;
-          const key = keys[j];
-          const softWrapped = softWrappedFlags[j];
-          const foldable = foldableFlags[j];
-          const bufferRow = bufferRows[j];
-          const screenRow = screenRows[j];
-
-          let className = 'line-number';
-          if (foldable) className = className + ' foldable';
-
-          const decorationsForRow = decorations[row - startRow];
-          if (decorationsForRow)
-            className = className + ' ' + decorationsForRow;
-
-          let number = null;
-          if (showLineNumbers) {
-            if (this.props.labelFn == null) {
-              number = softWrapped ? '•' : bufferRow + 1;
-              number =
-                NBSP_CHARACTER.repeat(maxDigits - number.length) + number;
-            } else {
-              number = this.props.labelFn({
-                bufferRow,
-                screenRow,
-                foldable,
-                softWrapped,
-                maxDigits
-              });
-            }
-          }
-
-          // We need to adjust the line number position to account for block
-          // decorations preceding the current row and following the preceding
-          // row. Note that we ignore the latter when the line number starts at
-          // the beginning of the tile, because the tile will already be
-          // positioned to take into account block decorations added after the
-          // last row of the previous tile.
-          let marginTop = rootComponent.heightForBlockDecorationsBeforeRow(row);
-          if (indexInTile > 0)
-            marginTop += rootComponent.heightForBlockDecorationsAfterRow(
-              row - 1
-            );
-
-          tileChildren[row - tileStartRow] = $(LineNumberComponent, {
-            key,
-            className,
-            width,
-            bufferRow,
-            screenRow,
-            number,
-            marginTop,
-            nodePool: this.nodePool
-          });
-        }
-
-        const tileTop = rootComponent.pixelPositionBeforeBlocksForRow(
-          tileStartRow
-        );
-        const tileBottom = rootComponent.pixelPositionBeforeBlocksForRow(
-          tileEndRow
-        );
-        const tileHeight = tileBottom - tileTop;
-        const tileWidth = width != null && width > 0 ? width + 'px' : '';
-
-        children[i] = $.div(
-          {
-            key: rootComponent.idsByTileStartRow.get(tileStartRow),
-            style: {
-              contain: 'layout style',
-              position: 'absolute',
-              top: 0,
-              height: tileHeight + 'px',
-              width: tileWidth,
-              transform: `translateY(${tileTop}px)`
-            }
-          },
-          ...tileChildren
-        );
-      }
-    }
-
-    let rootClassName = 'gutter line-numbers';
-    if (className) {
-      rootClassName += ' ' + className;
-    }
-
-    return $.div(
-      {
-        className: rootClassName,
-        attributes: { 'gutter-name': this.props.name },
-        style: {
-          position: 'relative',
-          height: ceilToPhysicalPixelBoundary(height) + 'px'
-        },
-        on: {
-          mousedown: this.didMouseDown,
-          mousemove: this.didMouseMove
-        }
-      },
-      $.div(
-        {
-          key: 'placeholder',
-          className: 'line-number dummy',
-          style: { visibility: 'hidden' }
-        },
-        showLineNumbers ? '0'.repeat(maxDigits) : null,
-        $.div({ className: 'icon-right' })
-      ),
-      children
-    );
-  }
-
-  shouldUpdate(newProps) {
-    const oldProps = this.props;
-
-    if (oldProps.showLineNumbers !== newProps.showLineNumbers) return true;
-    if (oldProps.height !== newProps.height) return true;
-    if (oldProps.width !== newProps.width) return true;
-    if (oldProps.lineHeight !== newProps.lineHeight) return true;
-    if (oldProps.startRow !== newProps.startRow) return true;
-    if (oldProps.endRow !== newProps.endRow) return true;
-    if (oldProps.rowsPerTile !== newProps.rowsPerTile) return true;
-    if (oldProps.maxDigits !== newProps.maxDigits) return true;
-    if (oldProps.labelFn !== newProps.labelFn) return true;
-    if (oldProps.className !== newProps.className) return true;
-    if (newProps.didMeasureVisibleBlockDecoration) return true;
-    if (!arraysEqual(oldProps.keys, newProps.keys)) return true;
-    if (!arraysEqual(oldProps.bufferRows, newProps.bufferRows)) return true;
-    if (!arraysEqual(oldProps.foldableFlags, newProps.foldableFlags))
-      return true;
-    if (!arraysEqual(oldProps.decorations, newProps.decorations)) return true;
-
-    let oldTileStartRow = oldProps.startRow;
-    let newTileStartRow = newProps.startRow;
-    while (
-      oldTileStartRow < oldProps.endRow ||
-      newTileStartRow < newProps.endRow
-    ) {
-      let oldTileBlockDecorations = oldProps.blockDecorations.get(
-        oldTileStartRow
-      );
-      let newTileBlockDecorations = newProps.blockDecorations.get(
-        newTileStartRow
-      );
-
-      if (oldTileBlockDecorations && newTileBlockDecorations) {
-        if (oldTileBlockDecorations.size !== newTileBlockDecorations.size)
-          return true;
-
-        let blockDecorationsChanged = false;
-
-        oldTileBlockDecorations.forEach((oldDecorations, screenLineId) => {
-          if (!blockDecorationsChanged) {
-            const newDecorations = newTileBlockDecorations.get(screenLineId);
-            blockDecorationsChanged =
-              newDecorations == null ||
-              !arraysEqual(oldDecorations, newDecorations);
-          }
-        });
-        if (blockDecorationsChanged) return true;
-
-        newTileBlockDecorations.forEach((newDecorations, screenLineId) => {
-          if (!blockDecorationsChanged) {
-            const oldDecorations = oldTileBlockDecorations.get(screenLineId);
-            blockDecorationsChanged = oldDecorations == null;
-          }
-        });
-        if (blockDecorationsChanged) return true;
-      } else if (oldTileBlockDecorations) {
-        return true;
-      } else if (newTileBlockDecorations) {
-        return true;
-      }
-
-      oldTileStartRow += oldProps.rowsPerTile;
-      newTileStartRow += newProps.rowsPerTile;
-    }
-
-    return false;
-  }
-
-  didMouseDown(event) {
-    if (this.props.onMouseDown == null) {
-      this.props.rootComponent.didMouseDownOnLineNumberGutter(event);
-    } else {
-      const { bufferRow, screenRow } = event.target.dataset;
-      this.props.onMouseDown({
-        bufferRow: parseInt(bufferRow, 10),
-        screenRow: parseInt(screenRow, 10),
-        domEvent: event
-      });
-    }
-  }
-
-  didMouseMove(event) {
-    if (this.props.onMouseMove != null) {
-      const { bufferRow, screenRow } = event.target.dataset;
-      this.props.onMouseMove({
-        bufferRow: parseInt(bufferRow, 10),
-        screenRow: parseInt(screenRow, 10),
-        domEvent: event
-      });
-    }
-  }
-}
-
-class LineNumberComponent {
-  constructor(props) {
-    const {
-      className,
-      width,
-      marginTop,
-      bufferRow,
-      screenRow,
-      number,
-      nodePool
-    } = props;
-    this.props = props;
-    const style = {};
-    if (width != null && width > 0) style.width = width + 'px';
-    if (marginTop != null && marginTop > 0) style.marginTop = marginTop + 'px';
-    this.element = nodePool.getElement('DIV', className, style);
-    this.element.dataset.bufferRow = bufferRow;
-    this.element.dataset.screenRow = screenRow;
-    if (number) this.element.appendChild(nodePool.getTextNode(number));
-    this.element.appendChild(nodePool.getElement('DIV', 'icon-right', null));
-  }
-
-  destroy() {
-    this.element.remove();
-    this.props.nodePool.release(this.element);
-  }
-
-  update(props) {
-    const {
-      nodePool,
-      className,
-      width,
-      marginTop,
-      bufferRow,
-      screenRow,
-      number
-    } = props;
-
-    if (this.props.bufferRow !== bufferRow)
-      this.element.dataset.bufferRow = bufferRow;
-    if (this.props.screenRow !== screenRow)
-      this.element.dataset.screenRow = screenRow;
-    if (this.props.className !== className) this.element.className = className;
-    if (this.props.width !== width) {
-      if (width != null && width > 0) {
-        this.element.style.width = width + 'px';
-      } else {
-        this.element.style.width = '';
-      }
-    }
-    if (this.props.marginTop !== marginTop) {
-      if (marginTop != null && marginTop > 0) {
-        this.element.style.marginTop = marginTop + 'px';
-      } else {
-        this.element.style.marginTop = '';
-      }
-    }
-
-    if (this.props.number !== number) {
-      if (this.props.number != null) {
-        const numberNode = this.element.firstChild;
-        numberNode.remove();
-        nodePool.release(numberNode);
-      }
-
-      if (number != null) {
-        this.element.insertBefore(
-          nodePool.getTextNode(number),
-          this.element.firstChild
-        );
-      }
-    }
-
-    this.props = props;
-  }
-}
-
-class CustomGutterComponent {
-  constructor(props) {
-    this.props = props;
-    this.element = this.props.element;
-    this.virtualNode = $.div(null);
-    this.virtualNode.domNode = this.element;
-    etch.updateSync(this);
-  }
-
-  update(props) {
-    this.props = props;
-    etch.updateSync(this);
-  }
-
-  destroy() {
-    etch.destroy(this);
-  }
-
-  render() {
-    let className = 'gutter';
-    if (this.props.className) {
-      className += ' ' + this.props.className;
-    }
-    return $.div(
-      {
-        className,
-        attributes: { 'gutter-name': this.props.name },
-        style: {
-          display: this.props.visible ? '' : 'none'
-        }
-      },
-      $.div(
-        {
-          className: 'custom-decorations',
-          style: { height: this.props.height + 'px' }
-        },
-        this.renderDecorations()
-      )
-    );
-  }
-
-  renderDecorations() {
-    if (!this.props.decorations) return null;
-
-    return this.props.decorations.map(({ className, element, top, height }) => {
-      return $(CustomGutterDecorationComponent, {
-        className,
-        element,
-        top,
-        height
-      });
-    });
-  }
-}
-
-class CustomGutterDecorationComponent {
-  constructor(props) {
-    this.props = props;
-    this.element = document.createElement('div');
-    const { top, height, className, element } = this.props;
-
-    this.element.style.position = 'absolute';
-    this.element.style.top = top + 'px';
-    this.element.style.height = height + 'px';
-    if (className != null) this.element.className = className;
-    if (element != null) {
-      this.element.appendChild(element);
-      element.style.height = height + 'px';
-    }
-  }
-
-  update(newProps) {
-    const oldProps = this.props;
-    this.props = newProps;
-
-    if (newProps.top !== oldProps.top)
-      this.element.style.top = newProps.top + 'px';
-    if (newProps.height !== oldProps.height) {
-      this.element.style.height = newProps.height + 'px';
-      if (newProps.element)
-        newProps.element.style.height = newProps.height + 'px';
-    }
-    if (newProps.className !== oldProps.className)
-      this.element.className = newProps.className || '';
-    if (newProps.element !== oldProps.element) {
-      if (this.element.firstChild) this.element.firstChild.remove();
-      if (newProps.element != null) {
-        this.element.appendChild(newProps.element);
-        newProps.element.style.height = newProps.height + 'px';
-      }
-    }
-  }
-}
-
-class CursorsAndInputComponent {
-  constructor(props) {
-    this.props = props;
-    etch.initialize(this);
-  }
-
-  update(props) {
-    if (props.measuredContent) {
-      this.props = props;
-      etch.updateSync(this);
-    }
-  }
-
-  updateCursorBlinkSync(cursorsBlinkedOff) {
-    this.props.cursorsBlinkedOff = cursorsBlinkedOff;
-    const className = this.getCursorsClassName();
-    this.refs.cursors.className = className;
-    this.virtualNode.props.className = className;
-  }
-
-  render() {
-    const {
-      lineHeight,
-      decorationsToRender,
-      scrollHeight,
-      scrollWidth
-    } = this.props;
-
-    const className = this.getCursorsClassName();
-    const cursorHeight = lineHeight + 'px';
-
-    const children = [this.renderHiddenInput()];
-    for (let i = 0; i < decorationsToRender.cursors.length; i++) {
-      const {
-        pixelLeft,
-        pixelTop,
-        pixelWidth,
-        className: extraCursorClassName,
-        style: extraCursorStyle
-      } = decorationsToRender.cursors[i];
-      let cursorClassName = 'cursor';
-      if (extraCursorClassName) cursorClassName += ' ' + extraCursorClassName;
-
-      const cursorStyle = {
-        height: cursorHeight,
-        width: Math.min(pixelWidth, scrollWidth - pixelLeft) + 'px',
-        transform: `translate(${pixelLeft}px, ${pixelTop}px)`
-      };
-      if (extraCursorStyle) Object.assign(cursorStyle, extraCursorStyle);
-
-      children.push(
-        $.div({
-          className: cursorClassName,
-          style: cursorStyle
-        })
-      );
-    }
-
-    return $.div(
-      {
-        key: 'cursors',
-        ref: 'cursors',
-        className,
-        style: {
-          position: 'absolute',
-          contain: 'strict',
-          zIndex: 1,
-          width: scrollWidth + 'px',
-          height: scrollHeight + 'px',
-          pointerEvents: 'none',
-          userSelect: 'none'
-        }
-      },
-      children
-    );
-  }
-
-  getCursorsClassName() {
-    return this.props.cursorsBlinkedOff ? 'cursors blink-off' : 'cursors';
-  }
-
-  renderHiddenInput() {
-    const {
-      lineHeight,
-      hiddenInputPosition,
-      didBlurHiddenInput,
-      didFocusHiddenInput,
-      didPaste,
-      didTextInput,
-      didKeydown,
-      didKeyup,
-      didKeypress,
-      didCompositionStart,
-      didCompositionUpdate,
-      didCompositionEnd,
-      tabIndex
-    } = this.props;
-
-    let top, left;
-    if (hiddenInputPosition) {
-      top = hiddenInputPosition.pixelTop;
-      left = hiddenInputPosition.pixelLeft;
-    } else {
-      top = 0;
-      left = 0;
-    }
-
-    return $.input({
-      ref: 'hiddenInput',
-      key: 'hiddenInput',
-      className: 'hidden-input',
-      on: {
-        blur: didBlurHiddenInput,
-        focus: didFocusHiddenInput,
-        paste: didPaste,
-        textInput: didTextInput,
-        keydown: didKeydown,
-        keyup: didKeyup,
-        keypress: didKeypress,
-        compositionstart: didCompositionStart,
-        compositionupdate: didCompositionUpdate,
-        compositionend: didCompositionEnd
-      },
-      tabIndex: tabIndex,
-      style: {
-        position: 'absolute',
-        width: '1px',
-        height: lineHeight + 'px',
-        top: top + 'px',
-        left: left + 'px',
-        opacity: 0,
-        padding: 0,
-        border: 0
-      }
-    });
-  }
-}
-
-class LinesTileComponent {
-  constructor(props) {
-    this.props = props;
-    etch.initialize(this);
-    this.createLines();
-    this.updateBlockDecorations({}, props);
-  }
-
-  update(newProps) {
-    if (this.shouldUpdate(newProps)) {
-      const oldProps = this.props;
-      this.props = newProps;
-      etch.updateSync(this);
-      if (!newProps.measuredContent) {
-        this.updateLines(oldProps, newProps);
-        this.updateBlockDecorations(oldProps, newProps);
-      }
-    }
-  }
-
-  destroy() {
-    for (let i = 0; i < this.lineComponents.length; i++) {
-      this.lineComponents[i].destroy();
-    }
-    this.lineComponents.length = 0;
-
-    return etch.destroy(this);
-  }
-
-  render() {
-    const { height, width, top } = this.props;
-
-    return $.div(
-      {
-        style: {
-          contain: 'layout style',
-          position: 'absolute',
-          height: height + 'px',
-          width: width + 'px',
-          transform: `translateY(${top}px)`
-        }
-      }
-      // Lines and block decorations will be manually inserted here for efficiency
-    );
-  }
-
-  createLines() {
-    const {
-      tileStartRow,
-      screenLines,
-      lineDecorations,
-      textDecorations,
-      nodePool,
-      displayLayer,
-      lineComponentsByScreenLineId
-    } = this.props;
-
-    this.lineComponents = [];
-    for (let i = 0, length = screenLines.length; i < length; i++) {
-      const component = new LineComponent({
-        screenLine: screenLines[i],
-        screenRow: tileStartRow + i,
-        lineDecoration: lineDecorations[i],
-        textDecorations: textDecorations[i],
-        displayLayer,
-        nodePool,
-        lineComponentsByScreenLineId
-      });
-      this.element.appendChild(component.element);
-      this.lineComponents.push(component);
-    }
-  }
-
-  updateLines(oldProps, newProps) {
-    const {
-      screenLines,
-      tileStartRow,
-      lineDecorations,
-      textDecorations,
-      nodePool,
-      displayLayer,
-      lineComponentsByScreenLineId
-    } = newProps;
-
-    const oldScreenLines = oldProps.screenLines;
-    const newScreenLines = screenLines;
-    const oldScreenLinesEndIndex = oldScreenLines.length;
-    const newScreenLinesEndIndex = newScreenLines.length;
-    let oldScreenLineIndex = 0;
-    let newScreenLineIndex = 0;
-    let lineComponentIndex = 0;
-
-    while (
-      oldScreenLineIndex < oldScreenLinesEndIndex ||
-      newScreenLineIndex < newScreenLinesEndIndex
-    ) {
-      const oldScreenLine = oldScreenLines[oldScreenLineIndex];
-      const newScreenLine = newScreenLines[newScreenLineIndex];
-
-      if (oldScreenLineIndex >= oldScreenLinesEndIndex) {
-        var newScreenLineComponent = new LineComponent({
-          screenLine: newScreenLine,
-          screenRow: tileStartRow + newScreenLineIndex,
-          lineDecoration: lineDecorations[newScreenLineIndex],
-          textDecorations: textDecorations[newScreenLineIndex],
-          displayLayer,
-          nodePool,
-          lineComponentsByScreenLineId
-        });
-        this.element.appendChild(newScreenLineComponent.element);
-        this.lineComponents.push(newScreenLineComponent);
-
-        newScreenLineIndex++;
-        lineComponentIndex++;
-      } else if (newScreenLineIndex >= newScreenLinesEndIndex) {
-        this.lineComponents[lineComponentIndex].destroy();
-        this.lineComponents.splice(lineComponentIndex, 1);
-
-        oldScreenLineIndex++;
-      } else if (oldScreenLine === newScreenLine) {
-        const lineComponent = this.lineComponents[lineComponentIndex];
-        lineComponent.update({
-          screenRow: tileStartRow + newScreenLineIndex,
-          lineDecoration: lineDecorations[newScreenLineIndex],
-          textDecorations: textDecorations[newScreenLineIndex]
-        });
-
-        oldScreenLineIndex++;
-        newScreenLineIndex++;
-        lineComponentIndex++;
-      } else {
-        const oldScreenLineIndexInNewScreenLines = newScreenLines.indexOf(
-          oldScreenLine
-        );
-        const newScreenLineIndexInOldScreenLines = oldScreenLines.indexOf(
-          newScreenLine
-        );
-        if (
-          newScreenLineIndex < oldScreenLineIndexInNewScreenLines &&
-          oldScreenLineIndexInNewScreenLines < newScreenLinesEndIndex
-        ) {
-          const newScreenLineComponents = [];
-          while (newScreenLineIndex < oldScreenLineIndexInNewScreenLines) {
-            // eslint-disable-next-line no-redeclare
-            var newScreenLineComponent = new LineComponent({
-              screenLine: newScreenLines[newScreenLineIndex],
-              screenRow: tileStartRow + newScreenLineIndex,
-              lineDecoration: lineDecorations[newScreenLineIndex],
-              textDecorations: textDecorations[newScreenLineIndex],
-              displayLayer,
-              nodePool,
-              lineComponentsByScreenLineId
-            });
-            this.element.insertBefore(
-              newScreenLineComponent.element,
-              this.getFirstElementForScreenLine(oldProps, oldScreenLine)
-            );
-            newScreenLineComponents.push(newScreenLineComponent);
-
-            newScreenLineIndex++;
-          }
-
-          this.lineComponents.splice(
-            lineComponentIndex,
-            0,
-            ...newScreenLineComponents
-          );
-          lineComponentIndex =
-            lineComponentIndex + newScreenLineComponents.length;
-        } else if (
-          oldScreenLineIndex < newScreenLineIndexInOldScreenLines &&
-          newScreenLineIndexInOldScreenLines < oldScreenLinesEndIndex
-        ) {
-          while (oldScreenLineIndex < newScreenLineIndexInOldScreenLines) {
-            this.lineComponents[lineComponentIndex].destroy();
-            this.lineComponents.splice(lineComponentIndex, 1);
-
-            oldScreenLineIndex++;
-          }
-        } else {
-          const oldScreenLineComponent = this.lineComponents[
-            lineComponentIndex
-          ];
-          // eslint-disable-next-line no-redeclare
-          var newScreenLineComponent = new LineComponent({
-            screenLine: newScreenLines[newScreenLineIndex],
-            screenRow: tileStartRow + newScreenLineIndex,
-            lineDecoration: lineDecorations[newScreenLineIndex],
-            textDecorations: textDecorations[newScreenLineIndex],
-            displayLayer,
-            nodePool,
-            lineComponentsByScreenLineId
-          });
-          this.element.insertBefore(
-            newScreenLineComponent.element,
-            oldScreenLineComponent.element
-          );
-          oldScreenLineComponent.destroy();
-          this.lineComponents[lineComponentIndex] = newScreenLineComponent;
-
-          oldScreenLineIndex++;
-          newScreenLineIndex++;
-          lineComponentIndex++;
-        }
-      }
-    }
-  }
-
-  getFirstElementForScreenLine(oldProps, screenLine) {
-    const blockDecorations = oldProps.blockDecorations
-      ? oldProps.blockDecorations.get(screenLine.id)
-      : null;
-    if (blockDecorations) {
-      const blockDecorationElementsBeforeOldScreenLine = [];
-      for (let i = 0; i < blockDecorations.length; i++) {
-        const decoration = blockDecorations[i];
-        if (decoration.position !== 'after') {
-          blockDecorationElementsBeforeOldScreenLine.push(
-            TextEditor.viewForItem(decoration.item)
-          );
-        }
-      }
-
-      for (
-        let i = 0;
-        i < blockDecorationElementsBeforeOldScreenLine.length;
-        i++
-      ) {
-        const blockDecorationElement =
-          blockDecorationElementsBeforeOldScreenLine[i];
-        if (
-          !blockDecorationElementsBeforeOldScreenLine.includes(
-            blockDecorationElement.previousSibling
-          )
-        ) {
-          return blockDecorationElement;
-        }
-      }
-    }
-
-    return oldProps.lineComponentsByScreenLineId.get(screenLine.id).element;
-  }
-
-  updateBlockDecorations(oldProps, newProps) {
-    const { blockDecorations, lineComponentsByScreenLineId } = newProps;
-
-    if (oldProps.blockDecorations) {
-      oldProps.blockDecorations.forEach((oldDecorations, screenLineId) => {
-        const newDecorations = newProps.blockDecorations
-          ? newProps.blockDecorations.get(screenLineId)
-          : null;
-        for (let i = 0; i < oldDecorations.length; i++) {
-          const oldDecoration = oldDecorations[i];
-          if (newDecorations && newDecorations.includes(oldDecoration))
-            continue;
-
-          const element = TextEditor.viewForItem(oldDecoration.item);
-          if (element.parentElement !== this.element) continue;
-
-          element.remove();
-        }
-      });
-    }
-
-    if (blockDecorations) {
-      blockDecorations.forEach((newDecorations, screenLineId) => {
-        const oldDecorations = oldProps.blockDecorations
-          ? oldProps.blockDecorations.get(screenLineId)
-          : null;
-        const lineNode = lineComponentsByScreenLineId.get(screenLineId).element;
-        let lastAfter = lineNode;
-
-        for (let i = 0; i < newDecorations.length; i++) {
-          const newDecoration = newDecorations[i];
-          const element = TextEditor.viewForItem(newDecoration.item);
-
-          if (oldDecorations && oldDecorations.includes(newDecoration)) {
-            if (newDecoration.position === 'after') {
-              lastAfter = element;
-            }
-            continue;
-          }
-
-          if (newDecoration.position === 'after') {
-            this.element.insertBefore(element, lastAfter.nextSibling);
-            lastAfter = element;
-          } else {
-            this.element.insertBefore(element, lineNode);
-          }
-        }
-      });
-    }
-  }
-
-  shouldUpdate(newProps) {
-    const oldProps = this.props;
-    if (oldProps.top !== newProps.top) return true;
-    if (oldProps.height !== newProps.height) return true;
-    if (oldProps.width !== newProps.width) return true;
-    if (oldProps.lineHeight !== newProps.lineHeight) return true;
-    if (oldProps.tileStartRow !== newProps.tileStartRow) return true;
-    if (oldProps.tileEndRow !== newProps.tileEndRow) return true;
-    if (!arraysEqual(oldProps.screenLines, newProps.screenLines)) return true;
-    if (!arraysEqual(oldProps.lineDecorations, newProps.lineDecorations))
-      return true;
-
-    if (oldProps.blockDecorations && newProps.blockDecorations) {
-      if (oldProps.blockDecorations.size !== newProps.blockDecorations.size)
-        return true;
-
-      let blockDecorationsChanged = false;
-
-      oldProps.blockDecorations.forEach((oldDecorations, screenLineId) => {
-        if (!blockDecorationsChanged) {
-          const newDecorations = newProps.blockDecorations.get(screenLineId);
-          blockDecorationsChanged =
-            newDecorations == null ||
-            !arraysEqual(oldDecorations, newDecorations);
-        }
-      });
-      if (blockDecorationsChanged) return true;
-
-      newProps.blockDecorations.forEach((newDecorations, screenLineId) => {
-        if (!blockDecorationsChanged) {
-          const oldDecorations = oldProps.blockDecorations.get(screenLineId);
-          blockDecorationsChanged = oldDecorations == null;
-        }
-      });
-      if (blockDecorationsChanged) return true;
-    } else if (oldProps.blockDecorations) {
-      return true;
-    } else if (newProps.blockDecorations) {
-      return true;
-    }
-
-    if (oldProps.textDecorations.length !== newProps.textDecorations.length)
-      return true;
-    for (let i = 0; i < oldProps.textDecorations.length; i++) {
-      if (
-        !textDecorationsEqual(
-          oldProps.textDecorations[i],
-          newProps.textDecorations[i]
-        )
-      )
-        return true;
-    }
-
-    return false;
-  }
-}
-
-class LineComponent {
-  constructor(props) {
-    const {
-      nodePool,
-      screenRow,
-      screenLine,
-      lineComponentsByScreenLineId,
-      offScreen
-    } = props;
-    this.props = props;
-    this.element = nodePool.getElement('DIV', this.buildClassName(), null);
-    this.element.dataset.screenRow = screenRow;
-    this.textNodes = [];
-
-    if (offScreen) {
-      this.element.style.position = 'absolute';
-      this.element.style.visibility = 'hidden';
-      this.element.dataset.offScreen = true;
-    }
-
-    this.appendContents();
-    lineComponentsByScreenLineId.set(screenLine.id, this);
-  }
-
-  update(newProps) {
-    if (this.props.lineDecoration !== newProps.lineDecoration) {
-      this.props.lineDecoration = newProps.lineDecoration;
-      this.element.className = this.buildClassName();
-    }
-
-    if (this.props.screenRow !== newProps.screenRow) {
-      this.props.screenRow = newProps.screenRow;
-      this.element.dataset.screenRow = newProps.screenRow;
-    }
-
-    if (
-      !textDecorationsEqual(
-        this.props.textDecorations,
-        newProps.textDecorations
-      )
-    ) {
-      this.props.textDecorations = newProps.textDecorations;
-      this.element.firstChild.remove();
-      this.appendContents();
-    }
-  }
-
-  destroy() {
-    const { nodePool, lineComponentsByScreenLineId, screenLine } = this.props;
-
-    if (lineComponentsByScreenLineId.get(screenLine.id) === this) {
-      lineComponentsByScreenLineId.delete(screenLine.id);
-    }
-
-    this.element.remove();
-    nodePool.release(this.element);
-  }
-
-  appendContents() {
-    const { displayLayer, nodePool, screenLine, textDecorations } = this.props;
-
-    this.textNodes.length = 0;
-
-    const { lineText, tags } = screenLine;
-    let openScopeNode = nodePool.getElement('SPAN', null, null);
-    this.element.appendChild(openScopeNode);
-
-    let decorationIndex = 0;
-    let column = 0;
-    let activeClassName = null;
-    let activeStyle = null;
-    let nextDecoration = textDecorations
-      ? textDecorations[decorationIndex]
-      : null;
-    if (nextDecoration && nextDecoration.column === 0) {
-      column = nextDecoration.column;
-      activeClassName = nextDecoration.className;
-      activeStyle = nextDecoration.style;
-      nextDecoration = textDecorations[++decorationIndex];
-    }
-
-    for (let i = 0; i < tags.length; i++) {
-      const tag = tags[i];
-      if (tag !== 0) {
-        if (displayLayer.isCloseTag(tag)) {
-          openScopeNode = openScopeNode.parentElement;
-        } else if (displayLayer.isOpenTag(tag)) {
-          const newScopeNode = nodePool.getElement(
-            'SPAN',
-            displayLayer.classNameForTag(tag),
-            null
-          );
-          openScopeNode.appendChild(newScopeNode);
-          openScopeNode = newScopeNode;
-        } else {
-          const nextTokenColumn = column + tag;
-          while (nextDecoration && nextDecoration.column <= nextTokenColumn) {
-            const text = lineText.substring(column, nextDecoration.column);
-            this.appendTextNode(
-              openScopeNode,
-              text,
-              activeClassName,
-              activeStyle
-            );
-            column = nextDecoration.column;
-            activeClassName = nextDecoration.className;
-            activeStyle = nextDecoration.style;
-            nextDecoration = textDecorations[++decorationIndex];
-          }
-
-          if (column < nextTokenColumn) {
-            const text = lineText.substring(column, nextTokenColumn);
-            this.appendTextNode(
-              openScopeNode,
-              text,
-              activeClassName,
-              activeStyle
-            );
-            column = nextTokenColumn;
-          }
-        }
-      }
-    }
-
-    if (column === 0) {
-      const textNode = nodePool.getTextNode(' ');
-      this.element.appendChild(textNode);
-      this.textNodes.push(textNode);
-    }
-
-    if (lineText.endsWith(displayLayer.foldCharacter)) {
-      // Insert a zero-width non-breaking whitespace, so that LinesYardstick can
-      // take the fold-marker::after pseudo-element into account during
-      // measurements when such marker is the last character on the line.
-      const textNode = nodePool.getTextNode(ZERO_WIDTH_NBSP_CHARACTER);
-      this.element.appendChild(textNode);
-      this.textNodes.push(textNode);
-    }
-  }
-
-  appendTextNode(openScopeNode, text, activeClassName, activeStyle) {
-    const { nodePool } = this.props;
-
-    if (activeClassName || activeStyle) {
-      const decorationNode = nodePool.getElement(
-        'SPAN',
-        activeClassName,
-        activeStyle
-      );
-      openScopeNode.appendChild(decorationNode);
-      openScopeNode = decorationNode;
-    }
-
-    const textNode = nodePool.getTextNode(text);
-    openScopeNode.appendChild(textNode);
-    this.textNodes.push(textNode);
-  }
-
-  buildClassName() {
-    const { lineDecoration } = this.props;
-    let className = 'line';
-    if (lineDecoration != null) className = className + ' ' + lineDecoration;
-    return className;
-  }
-}
-
-class HighlightsComponent {
-  constructor(props) {
-    this.props = {};
-    this.element = document.createElement('div');
-    this.element.className = 'highlights';
-    this.element.style.contain = 'strict';
-    this.element.style.position = 'absolute';
-    this.element.style.overflow = 'hidden';
-    this.element.style.userSelect = 'none';
-    this.highlightComponentsByKey = new Map();
-    this.update(props);
-  }
-
-  destroy() {
-    this.highlightComponentsByKey.forEach(highlightComponent => {
-      highlightComponent.destroy();
-    });
-    this.highlightComponentsByKey.clear();
-  }
-
-  update(newProps) {
-    if (this.shouldUpdate(newProps)) {
-      this.props = newProps;
-      const { height, width, lineHeight, highlightDecorations } = this.props;
-
-      this.element.style.height = height + 'px';
-      this.element.style.width = width + 'px';
-
-      const visibleHighlightDecorations = new Set();
-      if (highlightDecorations) {
-        for (let i = 0; i < highlightDecorations.length; i++) {
-          const highlightDecoration = highlightDecorations[i];
-          const highlightProps = Object.assign(
-            { lineHeight },
-            highlightDecorations[i]
-          );
-
-          let highlightComponent = this.highlightComponentsByKey.get(
-            highlightDecoration.key
-          );
-          if (highlightComponent) {
-            highlightComponent.update(highlightProps);
-          } else {
-            highlightComponent = new HighlightComponent(highlightProps);
-            this.element.appendChild(highlightComponent.element);
-            this.highlightComponentsByKey.set(
-              highlightDecoration.key,
-              highlightComponent
-            );
-          }
-
-          highlightDecorations[i].flashRequested = false;
-          visibleHighlightDecorations.add(highlightDecoration.key);
-        }
-      }
-
-      this.highlightComponentsByKey.forEach((highlightComponent, key) => {
-        if (!visibleHighlightDecorations.has(key)) {
-          highlightComponent.destroy();
-          this.highlightComponentsByKey.delete(key);
-        }
-      });
-    }
-  }
-
-  shouldUpdate(newProps) {
-    const oldProps = this.props;
-
-    if (!newProps.hasInitialMeasurements) return false;
-
-    if (oldProps.width !== newProps.width) return true;
-    if (oldProps.height !== newProps.height) return true;
-    if (oldProps.lineHeight !== newProps.lineHeight) return true;
-    if (!oldProps.highlightDecorations && newProps.highlightDecorations)
-      return true;
-    if (oldProps.highlightDecorations && !newProps.highlightDecorations)
-      return true;
-    if (oldProps.highlightDecorations && newProps.highlightDecorations) {
-      if (
-        oldProps.highlightDecorations.length !==
-        newProps.highlightDecorations.length
-      )
-        return true;
-
-      for (
-        let i = 0, length = oldProps.highlightDecorations.length;
-        i < length;
-        i++
-      ) {
-        const oldHighlight = oldProps.highlightDecorations[i];
-        const newHighlight = newProps.highlightDecorations[i];
-        if (oldHighlight.className !== newHighlight.className) return true;
-        if (newHighlight.flashRequested) return true;
-        if (oldHighlight.startPixelTop !== newHighlight.startPixelTop)
-          return true;
-        if (oldHighlight.startPixelLeft !== newHighlight.startPixelLeft)
-          return true;
-        if (oldHighlight.endPixelTop !== newHighlight.endPixelTop) return true;
-        if (oldHighlight.endPixelLeft !== newHighlight.endPixelLeft)
-          return true;
-        if (!oldHighlight.screenRange.isEqual(newHighlight.screenRange))
-          return true;
-      }
-    }
-  }
-}
-
-class HighlightComponent {
-  constructor(props) {
-    this.props = props;
-    etch.initialize(this);
-    if (this.props.flashRequested) this.performFlash();
-  }
-
-  destroy() {
-    if (this.timeoutsByClassName) {
-      this.timeoutsByClassName.forEach(timeout => {
-        window.clearTimeout(timeout);
-      });
-      this.timeoutsByClassName.clear();
-    }
-
-    return etch.destroy(this);
-  }
-
-  update(newProps) {
-    this.props = newProps;
-    etch.updateSync(this);
-    if (newProps.flashRequested) this.performFlash();
-  }
-
-  performFlash() {
-    const { flashClass, flashDuration } = this.props;
-    if (!this.timeoutsByClassName) this.timeoutsByClassName = new Map();
-
-    // If a flash of this class is already in progress, clear it early and
-    // flash again on the next frame to ensure CSS transitions apply to the
-    // second flash.
-    if (this.timeoutsByClassName.has(flashClass)) {
-      window.clearTimeout(this.timeoutsByClassName.get(flashClass));
-      this.timeoutsByClassName.delete(flashClass);
-      this.element.classList.remove(flashClass);
-      requestAnimationFrame(() => this.performFlash());
-    } else {
-      this.element.classList.add(flashClass);
-      this.timeoutsByClassName.set(
-        flashClass,
-        window.setTimeout(() => {
-          this.element.classList.remove(flashClass);
-        }, flashDuration)
-      );
-    }
-  }
-
-  render() {
-    const {
-      className,
-      screenRange,
-      lineHeight,
-      startPixelTop,
-      startPixelLeft,
-      endPixelTop,
-      endPixelLeft
-    } = this.props;
-    const regionClassName = 'region ' + className;
-
-    let children;
-    if (screenRange.start.row === screenRange.end.row) {
-      children = $.div({
-        className: regionClassName,
-        style: {
-          position: 'absolute',
-          boxSizing: 'border-box',
-          top: startPixelTop + 'px',
-          left: startPixelLeft + 'px',
-          width: endPixelLeft - startPixelLeft + 'px',
-          height: lineHeight + 'px'
-        }
-      });
-    } else {
-      children = [];
-      children.push(
-        $.div({
-          className: regionClassName,
-          style: {
-            position: 'absolute',
-            boxSizing: 'border-box',
-            top: startPixelTop + 'px',
-            left: startPixelLeft + 'px',
-            right: 0,
-            height: lineHeight + 'px'
-          }
-        })
-      );
-
-      if (screenRange.end.row - screenRange.start.row > 1) {
-        children.push(
-          $.div({
-            className: regionClassName,
-            style: {
-              position: 'absolute',
-              boxSizing: 'border-box',
-              top: startPixelTop + lineHeight + 'px',
-              left: 0,
-              right: 0,
-              height: endPixelTop - startPixelTop - lineHeight * 2 + 'px'
-            }
-          })
-        );
-      }
-
-      if (endPixelLeft > 0) {
-        children.push(
-          $.div({
-            className: regionClassName,
-            style: {
-              position: 'absolute',
-              boxSizing: 'border-box',
-              top: endPixelTop - lineHeight + 'px',
-              left: 0,
-              width: endPixelLeft + 'px',
-              height: lineHeight + 'px'
-            }
-          })
-        );
-      }
-    }
-
-    return $.div({ className: 'highlight ' + className }, children);
-  }
-}
-
-class OverlayComponent {
-  constructor(props) {
-    this.props = props;
-    this.element = document.createElement('atom-overlay');
-    if (this.props.className != null)
-      this.element.classList.add(this.props.className);
-    this.element.appendChild(this.props.element);
-    this.element.style.position = 'fixed';
-    this.element.style.zIndex = 4;
-    this.element.style.top = (this.props.pixelTop || 0) + 'px';
-    this.element.style.left = (this.props.pixelLeft || 0) + 'px';
-    this.currentContentRect = null;
-
-    // Synchronous DOM updates in response to resize events might trigger a
-    // "loop limit exceeded" error. We disconnect the observer before
-    // potentially mutating the DOM, and then reconnect it on the next tick.
-    // Note: ResizeObserver calls its callback when .observe is called
-    this.resizeObserver = new ResizeObserver(entries => {
-      const { contentRect } = entries[0];
-
-      if (
-        this.currentContentRect &&
-        (this.currentContentRect.width !== contentRect.width ||
-          this.currentContentRect.height !== contentRect.height)
-      ) {
-        this.resizeObserver.disconnect();
-        this.props.didResize(this);
-        process.nextTick(() => {
-          this.resizeObserver.observe(this.props.element);
-        });
-      }
-
-      this.currentContentRect = contentRect;
-    });
-    this.didAttach();
-    this.props.overlayComponents.add(this);
-  }
-
-  destroy() {
-    this.props.overlayComponents.delete(this);
-    this.didDetach();
-  }
-
-  getNextUpdatePromise() {
-    if (!this.nextUpdatePromise) {
-      this.nextUpdatePromise = new Promise(resolve => {
-        this.resolveNextUpdatePromise = () => {
-          this.nextUpdatePromise = null;
-          this.resolveNextUpdatePromise = null;
-          resolve();
-        };
-      });
-    }
-    return this.nextUpdatePromise;
-  }
-
-  update(newProps) {
-    const oldProps = this.props;
-    this.props = Object.assign({}, oldProps, newProps);
-    if (this.props.pixelTop != null)
-      this.element.style.top = this.props.pixelTop + 'px';
-    if (this.props.pixelLeft != null)
-      this.element.style.left = this.props.pixelLeft + 'px';
-    if (newProps.className !== oldProps.className) {
-      if (oldProps.className != null)
-        this.element.classList.remove(oldProps.className);
-      if (newProps.className != null)
-        this.element.classList.add(newProps.className);
-    }
-
-    if (this.resolveNextUpdatePromise) this.resolveNextUpdatePromise();
-  }
-
-  didAttach() {
-    this.resizeObserver.observe(this.props.element);
-  }
-
-  didDetach() {
-    this.resizeObserver.disconnect();
-  }
-}
 
 let rangeForMeasurement;
 function clientRectForRange(textNode, startIndex, endIndex) {
@@ -5046,166 +3411,4 @@ function clientRectForRange(textNode, startIndex, endIndex) {
   rangeForMeasurement.setStart(textNode, startIndex);
   rangeForMeasurement.setEnd(textNode, endIndex);
   return rangeForMeasurement.getBoundingClientRect();
-}
-
-function textDecorationsEqual(oldDecorations, newDecorations) {
-  if (!oldDecorations && newDecorations) return false;
-  if (oldDecorations && !newDecorations) return false;
-  if (oldDecorations && newDecorations) {
-    if (oldDecorations.length !== newDecorations.length) return false;
-    for (let j = 0; j < oldDecorations.length; j++) {
-      if (oldDecorations[j].column !== newDecorations[j].column) return false;
-      if (oldDecorations[j].className !== newDecorations[j].className)
-        return false;
-      if (!objectsEqual(oldDecorations[j].style, newDecorations[j].style))
-        return false;
-    }
-  }
-  return true;
-}
-
-function arraysEqual(a, b) {
-  if (a.length !== b.length) return false;
-  for (let i = 0, length = a.length; i < length; i++) {
-    if (a[i] !== b[i]) return false;
-  }
-  return true;
-}
-
-function objectsEqual(a, b) {
-  if (!a && b) return false;
-  if (a && !b) return false;
-  if (a && b) {
-    for (const key in a) {
-      if (a[key] !== b[key]) return false;
-    }
-    for (const key in b) {
-      if (a[key] !== b[key]) return false;
-    }
-  }
-  return true;
-}
-
-function constrainRangeToRows(range, startRow, endRow) {
-  if (range.start.row < startRow || range.end.row >= endRow) {
-    range = range.copy();
-    if (range.start.row < startRow) {
-      range.start.row = startRow;
-      range.start.column = 0;
-    }
-    if (range.end.row >= endRow) {
-      range.end.row = endRow;
-      range.end.column = 0;
-    }
-  }
-  return range;
-}
-
-function debounce(fn, wait) {
-  let timestamp, timeout;
-
-  function later() {
-    const last = Date.now() - timestamp;
-    if (last < wait && last >= 0) {
-      timeout = setTimeout(later, wait - last);
-    } else {
-      timeout = null;
-      fn();
-    }
-  }
-
-  return function() {
-    timestamp = Date.now();
-    if (!timeout) timeout = setTimeout(later, wait);
-  };
-}
-
-class NodePool {
-  constructor() {
-    this.elementsByType = {};
-    this.textNodes = [];
-  }
-
-  getElement(type, className, style) {
-    let element;
-    const elementsByDepth = this.elementsByType[type];
-    if (elementsByDepth) {
-      while (elementsByDepth.length > 0) {
-        const elements = elementsByDepth[elementsByDepth.length - 1];
-        if (elements && elements.length > 0) {
-          element = elements.pop();
-          if (elements.length === 0) elementsByDepth.pop();
-          break;
-        } else {
-          elementsByDepth.pop();
-        }
-      }
-    }
-
-    if (element) {
-      element.className = className || '';
-      element.attributeStyleMap.forEach((value, key) => {
-        if (!style || style[key] == null) element.style[key] = '';
-      });
-      if (style) Object.assign(element.style, style);
-      for (const key in element.dataset) delete element.dataset[key];
-      while (element.firstChild) element.firstChild.remove();
-      return element;
-    } else {
-      const newElement = document.createElement(type);
-      if (className) newElement.className = className;
-      if (style) Object.assign(newElement.style, style);
-      return newElement;
-    }
-  }
-
-  getTextNode(text) {
-    if (this.textNodes.length > 0) {
-      const node = this.textNodes.pop();
-      node.textContent = text;
-      return node;
-    } else {
-      return document.createTextNode(text);
-    }
-  }
-
-  release(node, depth = 0) {
-    const { nodeName } = node;
-    if (nodeName === '#text') {
-      this.textNodes.push(node);
-    } else {
-      let elementsByDepth = this.elementsByType[nodeName];
-      if (!elementsByDepth) {
-        elementsByDepth = [];
-        this.elementsByType[nodeName] = elementsByDepth;
-      }
-
-      let elements = elementsByDepth[depth];
-      if (!elements) {
-        elements = [];
-        elementsByDepth[depth] = elements;
-      }
-
-      elements.push(node);
-      for (let i = 0; i < node.childNodes.length; i++) {
-        this.release(node.childNodes[i], depth + 1);
-      }
-    }
-  }
-}
-
-function roundToPhysicalPixelBoundary(virtualPixelPosition) {
-  const virtualPixelsPerPhysicalPixel = 1 / window.devicePixelRatio;
-  return (
-    Math.round(virtualPixelPosition / virtualPixelsPerPhysicalPixel) *
-    virtualPixelsPerPhysicalPixel
-  );
-}
-
-function ceilToPhysicalPixelBoundary(virtualPixelPosition) {
-  const virtualPixelsPerPhysicalPixel = 1 / window.devicePixelRatio;
-  return (
-    Math.ceil(virtualPixelPosition / virtualPixelsPerPhysicalPixel) *
-    virtualPixelsPerPhysicalPixel
-  );
 }
